Move API into .Internal namespaces.
Move the following classes to .Internal.

-InMemoryTransaction
-InMemoryTransactionManager
-SqliteDatabaseProviderServices
-SqliteQuerySqlGenerator
-SqliteQuerySqlGeneratorFactory
-ISqlServerExpressionVisitor

Fix #5761
