+++ src/Microsoft.EntityFrameworkCore.InMemory/Microsoft.EntityFrameworkCore.InMemory.csproj
@@ -80,8 +80,8 @@
     <Compile Include="Query\Internal\InMemoryQueryModelVisitor.cs" />
     <Compile Include="Query\Internal\InMemoryQueryModelVisitorFactory.cs" />
     <Compile Include="Query\Internal\MaterializerFactory.cs" />
-    <Compile Include="Storage\InMemoryTransaction.cs" />
-    <Compile Include="Storage\InMemoryTransactionManager.cs" />
+    <Compile Include="Storage\Internal\InMemoryTransaction.cs" />
+    <Compile Include="Storage\Internal\InMemoryTransactionManager.cs" />
     <Compile Include="Storage\Internal\IInMemoryDatabase.cs" />
     <Compile Include="Storage\Internal\IInMemoryStore.cs" />
     <Compile Include="Storage\Internal\IInMemoryStoreSource.cs" />
