+++ src/Microsoft.EntityFrameworkCore.InMemory/Storage/Internal/InMemoryTransaction.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.EntityFrameworkCore.Storage
+namespace Microsoft.EntityFrameworkCore.Storage.Internal
 {
     public class InMemoryTransaction : IDbContextTransaction
     {
