+++ src/Microsoft.EntityFrameworkCore.InMemory/Storage/Internal/InMemoryTransactionManager.cs
@@ -9,7 +9,7 @@ using Microsoft.EntityFrameworkCore.Internal;
 using Microsoft.EntityFrameworkCore.Utilities;
 using Microsoft.Extensions.Logging;
 
-namespace Microsoft.EntityFrameworkCore.Storage
+namespace Microsoft.EntityFrameworkCore.Storage.Internal
 {
     public class InMemoryTransactionManager : IDbContextTransactionManager
     {
