+++ src/Microsoft.EntityFrameworkCore.SqlServer/Query/Sql/Internal/ISqlServerExpressionVisitor.cs
@@ -5,7 +5,7 @@ using System.Linq.Expressions;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Query.Expressions.Internal;
 
-namespace Microsoft.EntityFrameworkCore.Query.Sql
+namespace Microsoft.EntityFrameworkCore.Query.Sql.Internal
 {
     public interface ISqlServerExpressionVisitor
     {
