+++ src/Microsoft.EntityFrameworkCore.Sqlite/Query/Sql/Internal/SqliteQuerySqlGenerator.cs
@@ -7,7 +7,7 @@ using Microsoft.EntityFrameworkCore.Query.Expressions;
 using Microsoft.EntityFrameworkCore.Storage;
 using Microsoft.EntityFrameworkCore.Utilities;
 
-namespace Microsoft.EntityFrameworkCore.Query.Sql
+namespace Microsoft.EntityFrameworkCore.Query.Sql.Internal
 {
     public class SqliteQuerySqlGenerator : DefaultQuerySqlGenerator
     {
