+++ src/Microsoft.EntityFrameworkCore.Sqlite/Query/Sql/Internal/SqliteQuerySqlGeneratorFactory.cs
@@ -6,7 +6,7 @@ using Microsoft.EntityFrameworkCore.Query.Expressions;
 using Microsoft.EntityFrameworkCore.Storage;
 using Microsoft.EntityFrameworkCore.Utilities;
 
-namespace Microsoft.EntityFrameworkCore.Query.Sql
+namespace Microsoft.EntityFrameworkCore.Query.Sql.Internal
 {
     public class SqliteQuerySqlGeneratorFactory : QuerySqlGeneratorFactoryBase
     {
